+++ kubejs/server_scripts/recipe/removed_recipes.js
@@ -3,15 +3,15 @@ const RemovedRecipes = {
 }
 
 RemovedRecipes.ById = [
-    'spelunkery:gunpowder',
+    // 'spelunkery:gunpowder',
     'spelunkery:filling/portal_fluid',
     'spelunkery:emptying/portal_fluid',
     'spelunkery:mixing/portal_fluid',
     'spelunkery:raw_zinc',
     'spelunkery:raw_zinc_nugget',
-    'spelunkery:saltpeter_block',
-    'spelunkery:bone_meal',
-    'spelunkery:saltpeter_from_block',
+    // 'spelunkery:saltpeter_block',
+    // 'spelunkery:bone_meal',
+    // 'spelunkery:saltpeter_from_block',
 
     'immersive_aircraft:gyrodyne',